NETWORK ADMINISTRATION &  MANAGEMENT
1.	CONNECT THE NETWORK 
Based on the logical topology provided below, configure the network so all the devices are connected to each other. 
A. Connecting Devices 
i. 	Use DHCP for host IP allocation (The network IPs to be used should start from the number you have on the student list. (E.g. If you’re number 5 on the list your IP for the first LAN should be 5.0).  ii. 	For LAN 4 use a class B IP Address. iii. 	Use OSPF for traffic routing. 
	iv. 	In LAN 1 create two VLANS and use Router-on-a-stick for inter-vlan routing. 
 
2.	VPN CONFIGURATION 
Configure Site-to-Site VPN to provide a site-to-site IPSec connection between both VPN Gateways as shown in figure below.  
3.	ACL CONFIGURATION 
Create an ACL which prevents PC0 to access both Server0 and Laptop0. Apply it properly. 
Create an ACL which prevents PC5 from accessing to PC7 and PC6.